Add support for network states and handle loss of connection
Previously, a loss of network would cause a permanent error and
require restarting the upload. Now, an additional callback is
provided that allows the client to tell resumable whether or not
the network is connected. This enables the library to pause gracefully
when the network is disconnected, and resume its operations
when the network returns. Tested on iOS.